requirements.txt
django>=4.2,<5.0
groq>=0.4.0
pdfplumber>=0.10.0
python-docx>=1.1.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
Pillow>=10.0.0
requests>=2.31.0
reportlab>=4.0.0
